[FLINK-19412][python] Re-layer Python Operation Make it Possible to Provide only Python implementation
