+++ flink-python/pyflink/fn_execution/beam/beam_sdk_worker_main.py
@@ -18,10 +18,7 @@
 import sys
 
 # force to register the operations to SDK Harness
-try:
-    import pyflink.fn_execution.beam.beam_operations_fast
-except ImportError:
-    import pyflink.fn_execution.beam.beam_operations_slow
+import pyflink.fn_execution.beam.beam_operations # noqa # pylint: disable=unused-import
 
 # force to register the coders to SDK Harness
 import pyflink.fn_execution.beam.beam_coders # noqa # pylint: disable=unused-import
